Refactor authentication
Move to a pathless layout route.
