--- app/routes/auth.ts
@@ -1,16 +1,0 @@
-/*!
- * Copyright © 2022 United States Government as represented by the Administrator
- * of the National Aeronautics and Space Administration. No copyright is claimed
- * in the United States under Title 17, U.S. Code. All Other Rights Reserved.
- *
- * SPDX-License-Identifier: NASA-1.3
- */
-
-import type { LoaderFunction } from '@remix-run/node'
-import { authorize } from '~/lib/auth.server'
-export const loader: LoaderFunction = ({ request }) => authorize(request)
-
-// FIXME: workaround for https://github.com/remix-run/remix/issues/1828.
-// Resource routes (without browser code) don't get server pruning done.
-// Once fixed upstream, remove.
-export const meta = () => ({})
